Add Interaction Recognition
After learning how to send a message, the bot can recognize if an existing message has any interaction. And if it is the button we want, it takes specific action.

Also, I decided to use IDs from my .env file, it'll be safer I believe.

Next, a modal will pop up as a button interaction. :eyes: