Add Modal Interaction
I've set the modal form and interaction of it. Everything looks perfect so far. The next step will be saving this data into my Airtable database. Also categorizing the data is essential. I need to think about how to put the related data into the table. I don't want to do it manually or with my hand. :laughing: